sesion30/requirements.txt
django
pymongo
djangorestframework
dnspython
boto3
django-storages